Rename db::db to db::lib
