+++ web/src/db/mod.rs
@@ -1,2 +1,2 @@
-pub mod db;
+pub mod lib;
 pub mod station_record;
