+++ web/src/state.rs
@@ -1,10 +1,10 @@
-use crate::db::db;
+use crate::db::lib::{create_pool, DbPool};
 use restations_config::Config;
 use std::sync::Arc;
 
 /// The application's state that is available in [`crate::controllers`] and [`crate::middlewares`].
 pub struct AppState {
-    pub pool: db::DbPool,
+    pub pool: DbPool,
 }
 
 /// The application's state as it is shared across the application, e.g. in controllers and middlewares.
@@ -16,7 +16,7 @@ pub type SharedAppState = Arc<AppState>;
 ///
 /// This function creates an [`AppState`] based on the current [`restations_config::Config`].
 pub async fn init_app_state(_config: Config) -> AppState {
-    let pool = db::create_pool("stations.sqlite.db").await;
+    let pool = create_pool("stations.sqlite.db").await;
 
     AppState { pool }
 }
